Move utils and employee_scheduling into src
